+++ agent/src/main/java/com/teamscale/jacoco/agent/git_properties/InvalidGitPropertiesException.java
@@ -1,0 +1,14 @@
+package com.teamscale.jacoco.agent.git_properties;
+
+/**
+ * Thrown in case a git.properties file is found but it is malformed.
+ */
+public class InvalidGitPropertiesException extends Exception {
+	/*package*/ InvalidGitPropertiesException(String s, Throwable throwable) {
+		super(s, throwable);
+	}
+
+	/*package*/ InvalidGitPropertiesException(String s) {
+		super(s);
+	}
+}