+++ agent/src/main/java/com/teamscale/jacoco/agent/options/AgentOptions.java
@@ -3,36 +3,21 @@
 | Copyright (c) 2009-2018 CQSE GmbH                                        |
 |                                                                          |
 +-------------------------------------------------------------------------*/
-package com.teamscale.jacoco.agent;
+package com.teamscale.jacoco.agent.options;
 
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Date;
-import java.util.List;
-import java.util.Locale;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
-
-import org.conqat.lib.commons.assertion.CCSMAssert;
-import org.conqat.lib.commons.collections.PairList;
-import org.conqat.lib.commons.filesystem.FileSystemUtils;
-import org.jacoco.agent.rt.IAgent;
-import org.slf4j.Logger;
-
+import com.teamscale.client.FileSystemUtils;
 import com.teamscale.client.TeamscaleServer;
+import com.teamscale.jacoco.agent.Agent;
+import com.teamscale.jacoco.agent.AgentBase;
 import com.teamscale.jacoco.agent.commandline.Validator;
+import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocatingTransformer;
+import com.teamscale.jacoco.agent.git_properties.GitPropertiesLocator;
 import com.teamscale.jacoco.agent.store.IXmlStore;
+import com.teamscale.jacoco.agent.store.TimestampedFileStore;
 import com.teamscale.jacoco.agent.store.UploadStoreException;
-import com.teamscale.jacoco.agent.store.file.TimestampedFileStore;
 import com.teamscale.jacoco.agent.store.upload.azure.AzureFileStorageConfig;
 import com.teamscale.jacoco.agent.store.upload.azure.AzureFileStorageUploadStore;
+import com.teamscale.jacoco.agent.store.upload.delay.DelayedCommitDescriptorStore;
 import com.teamscale.jacoco.agent.store.upload.http.HttpUploadStore;
 import com.teamscale.jacoco.agent.store.upload.teamscale.TeamscaleUploadStore;
 import com.teamscale.jacoco.agent.testimpact.TestExecutionWriter;
@@ -41,8 +26,25 @@ import com.teamscale.jacoco.agent.util.AgentUtils;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
 import com.teamscale.report.EDuplicateClassFileBehavior;
 import com.teamscale.report.util.ClasspathWildcardIncludeFilter;
-
 import okhttp3.HttpUrl;
+import org.conqat.lib.commons.assertion.CCSMAssert;
+import org.conqat.lib.commons.collections.PairList;
+import org.jacoco.agent.rt.IAgent;
+import org.jacoco.core.runtime.WildcardMatcher;
+import org.slf4j.Logger;
+
+import java.io.File;
+import java.io.IOException;
+import java.lang.instrument.Instrumentation;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Date;
+import java.util.List;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
@@ -71,7 +73,7 @@ public class AgentOptions {
 	/**
 	 * The directory to write the XML traces to.
 	 */
-	/* package */ Path outputDirectory = null;
+	/* package */ Path outputDirectory = AgentUtils.getAgentDirectory().resolve("coverage");
 
 	/**
 	 * The URL to which to upload coverage zips.
@@ -108,12 +110,14 @@ public class AgentOptions {
 	/* package */ boolean shouldIgnoreDuplicateClassFiles = true;
 
 	/**
-	 * Include patterns to pass on to JaCoCo.
+	 * Include patterns for fully qualified class names to pass on to JaCoCo. See {@link WildcardMatcher} for the
+	 * pattern syntax. Individual patterns must be separated by ":".
 	 */
 	/* package */ String jacocoIncludes = null;
 
 	/**
-	 * Exclude patterns to pass on to JaCoCo.
+	 * Exclude patterns for fully qualified class names to pass on to JaCoCo. See {@link WildcardMatcher} for the
+	 * pattern syntax. Individual patterns must be separated by ":".
 	 */
 	/* package */ String jacocoExcludes = null;
 
@@ -150,7 +154,7 @@ public class AgentOptions {
 	}
 
 	/**
-	 * Validates the options and throws an exception if they're not valid.
+	 * Validates the options and returns a validator with all validation errors.
 	 */
 	/* package */ Validator getValidator() {
 		Validator validator = new Validator();
@@ -159,10 +163,7 @@ public class AgentOptions {
 			validator.isTrue(path.canRead(), "Path '" + path + "' is not readable");
 		}
 
-		validator.ensure(() -> {
-			CCSMAssert.isNotNull(outputDirectory, "You must specify an output directory");
-			FileSystemUtils.ensureDirectoryExists(outputDirectory.toFile());
-		});
+		validator.ensure(() -> FileSystemUtils.ensureDirectoryExists(outputDirectory.toFile()));
 
 		if (loggingConfig != null) {
 			validator.ensure(() -> {
@@ -177,8 +178,8 @@ public class AgentOptions {
 			});
 		}
 
-		validator.isTrue(!useTestwiseCoverageMode() || uploadUrl == null,
-				"'upload-url' option is " + "incompatible with Testwise coverage mode!");
+		validator.isTrue(!useTestwiseCoverageMode() || uploadUrl == null, "'upload-url' option is " +
+				"incompatible with Testwise coverage mode!");
 
 		validator.isFalse(uploadUrl == null && !additionalMetaDataFiles.isEmpty(),
 				"You specified additional meta data files to be uploaded but did not configure an upload URL");
@@ -186,17 +187,17 @@ public class AgentOptions {
 		validator.isTrue(teamscaleServer.hasAllRequiredFieldsNull() || teamscaleServer.hasAllRequiredFieldsSet(),
 				"You did provide some options prefixed with 'teamscale-', but not all required ones!");
 
-		validator.isTrue(
-				(azureFileStorageConfig.hasAllRequiredFieldsSet() || azureFileStorageConfig.hasAllRequiredFieldsNull()),
-				"If you want to upload data to an azure file storage you need to provide both "
-						+ "'azure-url' and 'azure-key' ");
+		validator.isTrue((azureFileStorageConfig.hasAllRequiredFieldsSet() || azureFileStorageConfig
+						.hasAllRequiredFieldsNull()),
+				"If you want to upload data to an Azure file storage you need to provide both " +
+						"'azure-url' and 'azure-key' ");
 
 		List<Boolean> configuredStores = Stream
 				.of(azureFileStorageConfig.hasAllRequiredFieldsSet(), teamscaleServer.hasAllRequiredFieldsSet(),
 						uploadUrl != null).filter(x -> x).collect(Collectors.toList());
 
-		validator.isTrue(configuredStores.size() <= 1, "You cannot configure multiple upload stores, "
-				+ "such as a teamscale instance, upload url or azure file storage");
+		validator.isTrue(configuredStores.size() <= 1, "You cannot configure multiple upload stores, " +
+				"such as a Teamscale instance, upload URL or Azure file storage");
 
 		return validator;
 	}
@@ -266,34 +267,59 @@ public class AgentOptions {
 	}
 
 	/**
-	 * Returns in instance of the agent that was configured. Either an agent
-	 * with interval based line-coverage dump or the HTTP server is used.
+	 * Returns in instance of the agent that was configured. Either an agent with interval based line-coverage dump or
+	 * the HTTP server is used.
 	 */
-	public AgentBase createAgent(IAgent agent) throws UploadStoreException, IOException {
+	public AgentBase createAgent(Instrumentation instrumentation, IAgent agent) throws UploadStoreException, IOException {
 		if (useTestwiseCoverageMode()) {
 			return new TestwiseCoverageAgent(this, agent, new TestExecutionWriter(getTempFile("test-execution", "json")));
 		} else {
-			return new Agent(this, agent);
+			return new Agent(this, instrumentation, agent);
 		}
 	}
 
 	/**
 	 * Creates the store to use for the coverage XMLs.
 	 */
-	public IXmlStore createStore() throws UploadStoreException {
+	public IXmlStore createStore(Instrumentation instrumentation) throws UploadStoreException {
 		TimestampedFileStore fileStore = new TimestampedFileStore(outputDirectory);
 		if (uploadUrl != null) {
 			return new HttpUploadStore(fileStore, uploadUrl, additionalMetaDataFiles);
 		}
 		if (teamscaleServer.hasAllRequiredFieldsSet()) {
+			if (teamscaleServer.commit == null) {
+				logger.info("You did not provide a commit to upload to directly, so the Agent will try and" +
+						" auto-detect it by searching all profiled Jar/War/Ear/... files for a git.properties file.");
+				return createDelayedTeamscaleStore(instrumentation, fileStore);
+			}
 			return new TeamscaleUploadStore(fileStore, teamscaleServer);
 		}
 
 		if (azureFileStorageConfig.hasAllRequiredFieldsSet()) {
-			return new AzureFileStorageUploadStore(fileStore, azureFileStorageConfig, additionalMetaDataFiles);
+			return new AzureFileStorageUploadStore(fileStore, azureFileStorageConfig,
+					additionalMetaDataFiles);
 		}
 
 		return fileStore;
+	}
+
+	private IXmlStore createDelayedTeamscaleStore(Instrumentation instrumentation, TimestampedFileStore fileStore)
+			throws UploadStoreException {
+
+		Path cacheDirectory;
+		try {
+			cacheDirectory = Files.createTempDirectory(outputDirectory, "upload-cache");
+		} catch (IOException e) {
+			throw new UploadStoreException(
+					"Unable to create temporary cache directory within " + outputDirectory, e);
+		}
+
+		TimestampedFileStore cacheStore = new TimestampedFileStore(cacheDirectory);
+		DelayedCommitDescriptorStore store = new DelayedCommitDescriptorStore(
+				commit -> new TeamscaleUploadStore(fileStore, teamscaleServer), cacheStore);
+		GitPropertiesLocator locator = new GitPropertiesLocator(store);
+		instrumentation.addTransformer(new GitPropertiesLocatingTransformer(locator, getLocationIncludeFilter()));
+		return store;
 	}
 
 	/**
@@ -318,7 +344,7 @@ public class AgentOptions {
 	/**
 	 * @see #shouldIgnoreDuplicateClassFiles
 	 */
-	public EDuplicateClassFileBehavior duplicateClassFileBehavior() {
+	public EDuplicateClassFileBehavior getDuplicateClassFileBehavior() {
 		if (shouldIgnoreDuplicateClassFiles) {
 			return EDuplicateClassFileBehavior.WARN;
 		} else {
@@ -332,16 +358,14 @@ public class AgentOptions {
 	}
 
 	/**
-	 * Returns the port at which the http server should listen for test
-	 * execution information or null if disabled.
+	 * Returns the port at which the http server should listen for test execution information or null if disabled.
 	 */
 	public Integer getHttpServerPort() {
 		return httpServerPort;
 	}
 
 	/**
-	 * Returns the name of the environment variable to read the test uniform
-	 * path from.
+	 * Returns the name of the environment variable to read the test uniform path from.
 	 */
 	public String getTestEnvironmentVariableName() {
 		return testEnvironmentVariable;
@@ -365,7 +389,7 @@ public class AgentOptions {
 	 * @see #jacocoIncludes
 	 * @see #jacocoExcludes
 	 */
-	public Predicate<String> getLocationIncludeFilter() {
+	public ClasspathWildcardIncludeFilter getLocationIncludeFilter() {
 		return new ClasspathWildcardIncludeFilter(jacocoIncludes, jacocoExcludes);
 	}