+++ agent/src/main/java/com/teamscale/jacoco/agent/store/upload/UploadStoreBase.java
@@ -2,10 +2,10 @@ package com.teamscale.jacoco.agent.store.upload;
 
 import com.teamscale.jacoco.agent.store.IXmlStore;
 import com.teamscale.jacoco.agent.store.UploadStoreException;
-import com.teamscale.jacoco.agent.store.file.TimestampedFileStore;
+import com.teamscale.jacoco.agent.store.TimestampedFileStore;
 import com.teamscale.jacoco.agent.util.Benchmark;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
-import eu.cqse.teamscale.client.HttpUtils;
+import com.teamscale.client.HttpUtils;
 import okhttp3.HttpUrl;
 import okhttp3.ResponseBody;
 import org.conqat.lib.commons.filesystem.FileSystemUtils;