+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageAgent.java
@@ -11,7 +11,7 @@ import org.jacoco.agent.rt.IAgent;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
 import com.teamscale.jacoco.agent.AgentBase;
-import com.teamscale.jacoco.agent.AgentOptions;
+import com.teamscale.jacoco.agent.options.AgentOptions;
 import com.teamscale.jacoco.agent.JacocoRuntimeController.DumpException;
 import com.teamscale.report.testwise.model.TestExecution;