+++ report-generator/src/main/java/com/teamscale/report/util/ClasspathWildcardIncludeFilter.java
@@ -5,13 +5,11 @@ import com.teamscale.client.StringUtils;
 import org.jacoco.core.runtime.WildcardMatcher;
 import org.jacoco.report.JavaNames;
 
-import java.util.function.Predicate;
-
 /***
  * Tests given class file paths against call name patterns.
  * E.g. "/some/file/path/test.jar@my/package/Test.class" matches "my/package/*" or "my/package/Test"
  */
-public class ClasspathWildcardIncludeFilter implements Predicate<String> {
+public class ClasspathWildcardIncludeFilter {
 
 	/**
 	 * Include patterns to apply during JaCoCo's traversal of class files. If null then everything is included.
@@ -26,8 +24,10 @@ public class ClasspathWildcardIncludeFilter implements Predicate<String> {
 	/**
 	 * Constructor.
 	 *
-	 * @param locationIncludeFilters Colon separated list of wildcard include patterns or null for no includes.
-	 * @param locationExcludeFilters Colon separated list of wildcard exclude patterns or null for no excludes.
+	 * @param locationIncludeFilters Colon separated list of wildcard include patterns for fully qualified class names
+	 *                               or null for no includes. See {@link WildcardMatcher} for the pattern syntax.
+	 * @param locationExcludeFilters Colon separated list of wildcard exclude patterns for fully qualified class names
+	 *                               or null for no excludes.See {@link WildcardMatcher} for the pattern syntax.
 	 */
 	public ClasspathWildcardIncludeFilter(String locationIncludeFilters, String locationExcludeFilters) {
 		if (locationIncludeFilters != null && !locationIncludeFilters.isEmpty()) {
@@ -38,8 +38,10 @@ public class ClasspathWildcardIncludeFilter implements Predicate<String> {
 		}
 	}
 
-	@Override
-	public boolean test(String path) {
+	/**
+	 * Tests if the given file path (e.g. "/some/file/path/test.jar@my/package/Test.class" or "org/mypackage/MyClass"
+	 */
+	public boolean isIncluded(String path) {
 		String className = getClassName(path);
 		// first check includes
 		if (locationIncludeFilters != null && !locationIncludeFilters.matches(className)) {
@@ -50,7 +52,8 @@ public class ClasspathWildcardIncludeFilter implements Predicate<String> {
 	}
 
 	/**
-	 * Returns the normalized class name of the given class file's path.
+	 * Returns the normalized class name of the given class file's path. I.e. turns something like
+	 * "/opt/deploy/some.jar@com/teamscale/Class.class" into something like "com.teamscale.Class".
 	 */
 	/* package */
 	static String getClassName(String path) {
@@ -60,7 +63,9 @@ public class ClasspathWildcardIncludeFilter implements Predicate<String> {
 		}
 
 		String pathInsideJar = parts[parts.length - 1];
-		String pathWithoutExtension = StringUtils.removeLastPart(pathInsideJar, '.');
-		return new JavaNames().getQualifiedClassName(pathWithoutExtension);
+		if (path.toLowerCase().endsWith(".class")) {
+			pathInsideJar = StringUtils.removeLastPart(pathInsideJar, '.');
+		}
+		return new JavaNames().getQualifiedClassName(pathInsideJar);
 	}
 }