+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 We use [semantic versioning][semver]
 
-# Next version
-- [fix] prevent NPE when trying to read manifest from Jar file
+# Next Release
+- [fix] WildFly won't start with agent attached
+- [feature] make `out` parameter optional with sensible fallback (subdirectory `coverage` in agent installation directory)
+- [feature] if no `teamscale-commit`, `teamscale-git-properties-jar` or `teamscale-commit-manifest-jar` is configured,
+  all loaded Jar/War/Ear/... files that contain profiled code are checked for a `git.properties` file. This allows auto-detecting a
+  `git.properties` file without any additional configuration. 
+
+# 15.1.1
+- [documentation] Configuration for SAP NetWeaver Java (>= 7.50) is now documented
 
 # 15.1.0
 - [feature] supplying a `class-dir` option is no longer mandatory
+- [feature] agent logs errors in case of empty coverage dumps (i.e. misconfigured agent)
+- [fix] prevent NPE when trying to read manifest from Jar file
 
 # 15.0.0
 - [feature] support for git.properties to supply commit information